+++ src/Filters/Filter.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace CodeZero\BrowserLocale\Filters;
+
+interface Filter
+{
+    /**
+     * Filter the locales.
+     *
+     * @param array $locales
+     *
+     * @return array
+     */
+    public function filter(array $locales);
+}