+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `BrowserLocale` will be documented in this file.
 
+## 3.0.0 (2018-04-01)
+
+- Use `Filter` classes to filter specific Locale info with a new `filter()` method
+- `getLocales()` no longer accepts an argument to filter results
+
 ## 2.0.0 (2018-03-30)
 
 - Rename main class & remove interface for simplicity